+++ src/apis/types/store.type.ts
@@ -103,14 +103,6 @@ export type getMenuListResponse = {
   ];
 };
 
-//리뷰 등록 api props type정의
-export type postReviewProps = {
-  consumerId: string | null;
-  storeId: number;
-  content: string;
-  score: number;
-};
-
 //가게 사진 가져오기 api response type정의
 export type getImgResponse = {
   data: [
@@ -119,37 +111,6 @@ export type getImgResponse = {
       url: string;
     }
   ];
-};
-
-//가게 리뷰리스트 가져오기 api response type정의
-export type getStoreReviewResponse = {
-  data: [
-    {
-      reviewId: number;
-      content: string;
-      score: number;
-      createdDate: string;
-      consumerName: string;
-      consumerId: number;
-      consumerAvgScore: number;
-    }
-  ];
-  currentPage: number;
-  totalPages: number;
-  totalElements: number;
-  pageSize: number;
-  isFirst: boolean;
-  isLast: boolean;
-};
-
-//가게 댓글 대댓글 달기 api response type정의
-export type postStoreReviewReplyResponse = {
-  data: {
-    commentId: number;
-    content: string;
-    isOwner: boolean;
-    createDate: string;
-  };
 };
 
 //카테고리 리스트 가져오기 api response type정의
